refactor: move duplicate slide to store
